+++ client/src/components/modals/leo/nameSearchModal.tsx
@@ -44,6 +44,7 @@ const NameSearchModal: React.FC<Props> = ({ search, names, searchName, saveNote,
 
   function onSubmit(e: React.FormEvent) {
     e.preventDefault();
+    if (!name.value) return;
 
     searchName(name.value);
   }
@@ -456,13 +457,13 @@ const NameSearchModal: React.FC<Props> = ({ search, names, searchName, saveNote,
                   {/* weapons */}
                   <div className="mt-3" id="weapons">
                     <h5>
-                      {lang.citizen.vehicle.reged_vehicle} ({search.weapons.length})
+                      {lang.citizen.weapon.reged_weapons} ({search.weapons.length})
                     </h5>
 
                     <ul style={{ maxHeight: "20rem" }} className="list-group overflow-auto">
                       {!search.weapons[0] ? (
                         <li className="list-group-item border-dark text-dark">
-                          {lang.record.no_vehicles}
+                          {lang.record.no_weapons}
                         </li>
                       ) : (
                         search.weapons.map((weapon: Weapon, idx: number) => {
@@ -526,7 +527,7 @@ const NameSearchModal: React.FC<Props> = ({ search, names, searchName, saveNote,
           <button type="button" className="btn btn-secondary" data-bs-dismiss="modal">
             {lang.global.close}
           </button>
-          <button type="submit" disabled={name === ""} className="btn btn-primary">
+          <button type="submit" disabled={!name.value} className="btn btn-primary">
             {lang.global.search}
           </button>
         </div>