+++ client/src/lib/actions/admin.ts
@@ -24,6 +24,7 @@ import {
   DELETE_10_CODE,
   UPDATE_10_CODE,
   DELETE_PENAL_CODE,
+  REMOVE_USER,
 } from "../types";
 import lang from "../../language.json";
 import Logger from "../Logger";
@@ -152,6 +153,21 @@ export const unBanMember = (id: string) => async (dispatch: Dispatch<IDispatch>)
   }
 };
 
+export const removeUser = (id: string) => async (dispatch: Dispatch<IDispatch>) => {
+  try {
+    const res = await handleRequest(`/admin/management/members/remove/${id}`, "PUT");
+
+    if (isSuccess(res)) {
+      dispatch({
+        type: REMOVE_USER,
+      });
+      return (window.location.href = "/admin/manage/members");
+    }
+  } catch (e) {
+    Logger.error(UN_BAN_MEMBER, e);
+  }
+};
+
 export const acceptUser = (id: string) => async (dispatch: Dispatch<IDispatch>) => {
   try {
     const res = await handleRequest(`/admin/management/members/accept/${id}`, "PUT");
@@ -263,8 +279,8 @@ export const updateCadSettings = (data: {
   cad_name: string;
   whitelisted: string;
   tow_whitelisted: string;
-  company_whitelisted: string;
   webhook_url: string;
+  plate_length: number;
 }) => async (dispatch: Dispatch<IDispatch>) => {
   try {
     const res = await handleRequest("/admin/management/cad-settings", "PUT", data);