+++ client/src/pages/index.tsx
@@ -48,9 +48,6 @@ const Credits: React.FC = () => {
   return (
     <div className="fixed-bottom text-light bg-dark border-dark px-3 pt-3 d-flex justify-content-between">
       <p>
-        Feel Free to support me <a href="https://www.paypal.me/caspertheghosty">here</a>
-      </p>
-      <p>
         <a href="https://caspertheghost.me">CasperTheGhost</a> | Thanks for choosing SnailyCAD!
       </p>
     </div>