+++ server/src/routes/citizen/index.ts
@@ -70,7 +70,7 @@ router.post("/", useAuth, async (req: IRequest, res: Response) => {
     }
 
     const query =
-      "INSERT INTO `citizens` (`id`, `full_name`, `user_id`, `birth`, `gender`, `ethnicity`, `hair_color`, `eye_color`, `address`, `height`, `weight`, `dmv`, `fire_license`, `pilot_license`, `ccw`, `business`, `business_id`, `rank`, `vehicle_reg`, `posts`, `image_id`, `b_status`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)";
+      "INSERT INTO `citizens` (`id`, `full_name`, `user_id`, `birth`, `gender`, `ethnicity`, `hair_color`, `eye_color`, `address`, `height`, `weight`, `dmv`, `fire_license`, `pilot_license`, `ccw`, `business`, `business_id`, `rank`, `vehicle_reg`, `posts`, `image_id`, `b_status`, `note`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)";
 
     const id = uuidv4();
     try {
@@ -97,6 +97,7 @@ router.post("/", useAuth, async (req: IRequest, res: Response) => {
         true /* posts */,
         imageId /* image_id */,
         "" /* b_status */,
+        "" /* note */,
       ]);
     } catch (e) {
       Logger.error("CREATE_CITIZEN_ERROR", e);