+++ CHANGELOG.md
@@ -2,9 +2,21 @@
 
 **If you need help updating/installing or have found any bugs, please join [our Discord server](https://discord.gg/eGnrPqEH7U) or open a [GitHub issue here](https://github.com/Dev-CasperTheGhost/snaily-cadv3/issues/new/choose)**
 
+## 1.3.1
+
+- New: Signal 100
+- New: Able to change length of plate (Default: 8)
+- New: Able to delete users from the CAD (At the bottom of manage member page)
+- New: A bleet now shows the username of who created the bleet
+- Minor: Fixed bug admins not able to delete a bleet
+- Minor: Upgraded to [bootstrap v5 beta 2](https://getbootstrap.com/docs/5.0/getting-started/introduction/)
+- Minor: Updated screenshots
+- Minor: Fixed issue with creating a new citizen
+- Minor: Fixed duplicate naming in NameSearchModal
+
 ## 1.3.0
 
-- Add version check in client side
+- Minor: Add version check in client side
 - Major: Server side rendering (No more CORS errors!)
 - Major: Customize penal codes and 10 codes